+++ Game of Life/game_structure/settings.cpp
@@ -1,6 +1,6 @@
 #include "game_structure.h"
 
-void Game_Board::light_spaceship(size_t line_beg, size_t col_beg){
+void Game_Board::light_spaceship(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg >= game::BOARD_SIZE - 2 or col_beg >= game::BOARD_SIZE - 4){
 	   int erreur(0);
 	   throw erreur;
@@ -9,7 +9,7 @@ void Game_Board::light_spaceship(size_t line_beg, size_t col_beg){
 	 board[line_beg][col_beg + 4]->live(); board[line_beg + 1][col_beg + 4]->live(); board[line_beg + 2][col_beg + 3]->live(); board[line_beg + 2][col_beg]->live(); board[line_beg][col_beg]->live();
 }
 
-void Game_Board::square(size_t line_beg, size_t col_beg){
+void Game_Board::square(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg >= game::BOARD_SIZE - 1 or line_beg < 0 or col_beg >= game::BOARD_SIZE - 1 or col_beg < 0){
 		int erreur(0);
 		throw erreur;
@@ -17,7 +17,7 @@ void Game_Board::square(size_t line_beg, size_t col_beg){
 	board[line_beg][col_beg]->live(); board[line_beg + 1][col_beg]->live(); board[line_beg][col_beg + 1]->live(); board[line_beg + 1][col_beg + 1]->live();
 }
 
-void Game_Board::toad(size_t line_beg, size_t col_beg){
+void Game_Board::toad(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg < 1 or line_beg >= game::BOARD_SIZE + 2 or col_beg < 1 or col_beg >= game::BOARD_SIZE - 2) {
 		int erreur(0);
 		throw erreur;
@@ -26,7 +26,7 @@ void Game_Board::toad(size_t line_beg, size_t col_beg){
 	board[line_beg][col_beg + 2]->live(); board[line_beg + 1][col_beg + 2]->live();
 }
 
-void Game_Board::pulsar(size_t line_beg, size_t col_beg){
+void Game_Board::pulsar(unsigned short int line_beg, unsigned short int col_beg){
 	try {
 		for (int i(0); i <= 12; ++i){
 			if (i == 0 or i == 5 or i == 7 or i == 12){
@@ -42,7 +42,7 @@ void Game_Board::pulsar(size_t line_beg, size_t col_beg){
 	}
 }
 
-void Game_Board::beacon(size_t line_beg, size_t col_beg){
+void Game_Board::beacon(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg < 0 or line_beg >= game::BOARD_SIZE - 2 or col_beg < 0 or col_beg >= game::BOARD_SIZE - 2){
 		int erreur(0);
 		throw erreur;
@@ -51,7 +51,7 @@ void Game_Board::beacon(size_t line_beg, size_t col_beg){
 	board[line_beg + 1][col_beg + 2]->live(); board[line_beg + 2][col_beg + 1]->live(); board[line_beg + 2][col_beg + 2]->live();
 }
 
-void Game_Board::glider(size_t line_beg, size_t col_beg){
+void Game_Board::glider(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg < 0 or line_beg >= game::BOARD_SIZE - 2 or col_beg < 0 or col_beg >= game::BOARD_SIZE - 2){
 		int erreur(0);
 		throw erreur;
@@ -59,7 +59,7 @@ void Game_Board::glider(size_t line_beg, size_t col_beg){
 	board[line_beg][col_beg]->live(); board[line_beg][col_beg + 2]->live(); board[line_beg + 1][col_beg + 1]->live(); board[line_beg + 1][col_beg + 2]->live(); board[line_beg + 2][col_beg + 1]->live();
 }
 
-void Game_Board::mid_spaceship(size_t line_beg, size_t col_beg){
+void Game_Board::mid_spaceship(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg < 0 or line_beg >= game::BOARD_SIZE - 4 or col_beg < 0 or col_beg >= game::BOARD_SIZE - 5){
 		int erreur(0);
 		throw erreur;
@@ -69,7 +69,7 @@ void Game_Board::mid_spaceship(size_t line_beg, size_t col_beg){
 	board[line_beg + 3][col_beg + 5]->live(); board[line_beg + 4][col_beg + 3]->live(); 
 }
 
-void Game_Board::larger_blinker(size_t line_beg, size_t col_beg){
+void Game_Board::larger_blinker(unsigned short int line_beg, unsigned short int col_beg){
 	try {
 		drawLine(3, line_beg, col_beg);
 		drawCol(3, line_beg -1, col_beg + 1);
@@ -79,7 +79,7 @@ void Game_Board::larger_blinker(size_t line_beg, size_t col_beg){
 	}
 }
 
-void Game_Board::acorn(size_t line_beg, size_t col_beg){
+void Game_Board::acorn(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg < 0 or line_beg >= game::BOARD_SIZE - 1 or col_beg < 0 or col_beg >= game::BOARD_SIZE - 3){
 		int err(0);
 		throw err;
@@ -94,7 +94,7 @@ void Game_Board::acorn(size_t line_beg, size_t col_beg){
 	board[line_beg][col_beg + 1]->live(); board[line_beg + 1][col_beg + 3]->live();
 }
 
-void Game_Board::die_hard(size_t line_beg, size_t col_beg){
+void Game_Board::die_hard(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg < 0 or line_beg >= game::BOARD_SIZE - 2 or col_beg < 0 or col_beg >= game::BOARD_SIZE - 6){
 		int err(0);
 		throw err;
@@ -109,7 +109,7 @@ void Game_Board::die_hard(size_t line_beg, size_t col_beg){
 	board[line_beg + 2][col_beg + 1]->live(); board[line_beg][col_beg + 6]->live();
 }
 
-void Game_Board::R_pentomino(size_t line_beg, size_t col_beg){
+void Game_Board::R_pentomino(unsigned short int line_beg, unsigned short int col_beg){
 	if (line_beg < 0 or line_beg >= game::BOARD_SIZE - 1 or col_beg < 0 or col_beg >= game::BOARD_SIZE - 2){
 		int err(0);
 		throw err;
@@ -123,7 +123,7 @@ void Game_Board::R_pentomino(size_t line_beg, size_t col_beg){
 	board[line_beg + 1][col_beg]->live(); board[line_beg][col_beg + 2]->live();
 }
 
-void Game_Board::Pi_heptomino(size_t line_beg, size_t col_beg){
+void Game_Board::Pi_heptomino(unsigned short int line_beg, unsigned short int col_beg){
 	try {
 		drawCol(3, line_beg, col_beg);
 		drawCol(3, line_beg, col_beg + 2);
@@ -135,8 +135,8 @@ void Game_Board::Pi_heptomino(size_t line_beg, size_t col_beg){
 }
 
 void Game_Board::nice_board(void){
-	for (size_t i(2); i < game::BOARD_SIZE - 2; ++i){
-		for (size_t j(2); j < game::BOARD_SIZE - 2; ++j){
+	for (unsigned short int i(2); i < game::BOARD_SIZE - 2; ++i){
+		for (unsigned short int j(2); j < game::BOARD_SIZE - 2; ++j){
 			if (i%2 == 0 and j%2 == 0){
 				board[i][j]->live();
 			}
@@ -145,8 +145,8 @@ void Game_Board::nice_board(void){
 }
 
 void Game_Board::checker_board(void){
-	for (size_t i(2); i < game::BOARD_SIZE - 1; ++i){
-		for (size_t j(2); j < game::BOARD_SIZE - 1; ++j){
+	for (unsigned short int i(2); i < game::BOARD_SIZE - 1; ++i){
+		for (unsigned short int j(2); j < game::BOARD_SIZE - 1; ++j){
 			if (i%2 != j%2){
 				board[i][j]->live();
 			}
@@ -154,7 +154,7 @@ void Game_Board::checker_board(void){
 	}
 }
 
-void Game_Board::gosper_glider_gun(size_t line_beg, size_t col_beg){
+void Game_Board::gosper_glider_gun(unsigned short int line_beg, unsigned short int col_beg){
 	square(line_beg, col_beg);
 	drawCol(3,line_beg, col_beg+10);
 	board[line_beg -1][col_beg+11]->live();
